requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
distro==1.9.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.33
idna==3.6
nodeenv==1.8.0
openai==1.11.1
pdf2image==1.17.0
platformdirs==4.2.0
pre-commit==3.6.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
ruff==0.2.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
